Renaming implicit gemm header
